[Feat] Game 씬 배치만, 일부 스크립트 파일 이동있음 sorting layer 3개 사용

+++ Assets/3.Script/Object/Object_FallBoxController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Object_FallBoxController : MonoBehaviour {
-    public int FallBoxRemainTime;
+    public float FallBoxRemainTime;
 
     private RectTransform rectTransform;
     private Rigidbody2D fallBoxRigidyBody;
